fixed using the controllers
