+++ routes/apiRoutes.js
@@ -4,70 +4,17 @@ const db = require("../models");
 const express = require("express");
 
 const apiRouter = express.Router();//for api routes
-const Helper = require("./utils/helper");
+const APIController = require("./controller/apiController");
 
 //TODO: use async/await
 // Get all data from current user
-apiRouter.get("/alldata", APIController);
+apiRouter.get("/alldata", APIController.getAllDataFromUser);
 
 //post a new stool log for user
-apiRouter.post("/stool", (req, res, next)=> {
-    const userId = Helper.checkAuth(req.user);
-    if(userId){
-        if (req.body && req.body.score && !isNaN(parseInt(req.body.score))) {
-            db.User.findOne({
-                where: {
-                    id: userId
-                }
-            }).then(function (userResult) {
-                userResult.createStool({
-                    score: req.body.score,
-                    time: Helper.convertTime(req.body),
-                    comment: req.body.comment ? req.body.comment : null//TODO:sanitize the input
-                }).then(stool => {
-                    res.json(stool);
-                }).catch(err=>{
-                    next(err);
-                });
-            }).catch(err=>{
-                next(err);
-            });
-        } else {
-            res.status(400).send("invalid input");
-        }
-    }else{
-        next(401);
-    }
-});
+apiRouter.post("/stool", APIController.postStoolDataForUser);
 
 
-apiRouter.post("/water", (req, res, next)=> {
-    const userId = Helper.checkAuth(req.user);
-    if(userId){
-        if(req.body && req.body.intake && !isNaN(parseInt(req.body.intake))){
-            db.User.findOne({
-                where:{
-                    id:userId
-                }
-            }).then(function(userResult) {
-                userResult.createWater({
-                    intake:req.body.intake,
-                    time:Helper.convertTime(req.body)
-                }).then(water=>{
-                    res.json(water);
-                }).catch(err=>{
-                    next(err);
-                });
-            }).catch(err=>{
-                next(err);
-            });
-        }else{
-            res.status(400).send("invalid input");
-        }
-    }else{
-        next(401);
-    }
-});
+apiRouter.post("/water", APIController.postWaterDataForUser);
 
 // TODO: Delete an user by id
 apiRouter.delete("/api/users/:id", (req, res)=> {
